Rename AgentDatabase to AgentStore
